OpenMonitor/django_app/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
requests
